Remove deprecated code from certificates controller
