+++ pkg/controller/certificates/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "certificate_request.go",
         "checks.go",
         "controller.go",
         "sync.go",
+        "util.go",
     ],
     importpath = "github.com/jetstack/cert-manager/pkg/controller/certificates",
     visibility = ["//visibility:public"],
@@ -17,8 +17,6 @@ go_library(
         "//pkg/client/listers/certmanager/v1alpha1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/feature:go_default_library",
-        "//pkg/internal/apis/certmanager/validation:go_default_library",
-        "//pkg/issuer:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/metrics:go_default_library",
         "//pkg/scheduler:go_default_library",
@@ -33,7 +31,6 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/errors:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
         "//vendor/k8s.io/client-go/listers/core/v1:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
@@ -59,26 +56,17 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "certificate_request_test.go",
-        "sync_test.go",
-    ],
+    srcs = ["sync_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//pkg/apis/certmanager/v1alpha1:go_default_library",
         "//pkg/controller/test:go_default_library",
-        "//pkg/feature:go_default_library",
-        "//pkg/issuer:go_default_library",
-        "//pkg/issuer/fake:go_default_library",
-        "//pkg/issuer/selfsigned:go_default_library",
-        "//pkg/util/feature:go_default_library",
         "//pkg/util/pki:go_default_library",
         "//test/unit/gen:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/client-go/testing:go_default_library",
-        "//vendor/k8s.io/component-base/featuregate/testing:go_default_library",
         "//vendor/k8s.io/utils/clock/testing:go_default_library",
     ],
 )
