+++ pkg/controller/certificates/sync.go
@@ -18,586 +18,809 @@ package certificates
 
 import (
 	"context"
-	"crypto"
+	"crypto/ecdsa"
+	"crypto/rsa"
 	"crypto/x509"
 	"encoding/json"
+	"encoding/pem"
 	"fmt"
-	"math/big"
+	"hash/fnv"
 	"reflect"
 	"strings"
 	"time"
 
-	"github.com/kr/pretty"
-	corev1 "k8s.io/api/core/v1"
-	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	utilerrors "k8s.io/apimachinery/pkg/util/errors"
-	corelisters "k8s.io/client-go/listers/core/v1"
-	"k8s.io/client-go/tools/cache"
-
+	"github.com/go-logr/logr"
 	apiutil "github.com/jetstack/cert-manager/pkg/api/util"
-	"github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
-	cmclient "github.com/jetstack/cert-manager/pkg/client/clientset/versioned"
-	"github.com/jetstack/cert-manager/pkg/feature"
-	"github.com/jetstack/cert-manager/pkg/internal/apis/certmanager/validation"
-	"github.com/jetstack/cert-manager/pkg/issuer"
+	cmapi "github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
+	cmlisters "github.com/jetstack/cert-manager/pkg/client/listers/certmanager/v1alpha1"
 	logf "github.com/jetstack/cert-manager/pkg/logs"
 	"github.com/jetstack/cert-manager/pkg/metrics"
-	"github.com/jetstack/cert-manager/pkg/util"
 	"github.com/jetstack/cert-manager/pkg/util/errors"
-	utilfeature "github.com/jetstack/cert-manager/pkg/util/feature"
 	"github.com/jetstack/cert-manager/pkg/util/kube"
 	"github.com/jetstack/cert-manager/pkg/util/pki"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 )
 
-const (
-	errorIssuerNotFound      = "IssuerNotFound"
-	errorIssuerNotReady      = "IssuerNotReady"
-	errorIssuerInit          = "IssuerInitError"
-	errorSavingCertificate   = "SaveCertError"
-	errorConfig              = "ConfigError"
-	errorDuplicateSecretName = "DuplicateSecretNameError"
-
-	reasonIssuingCertificate  = "IssueCert"
-	reasonRenewingCertificate = "RenewCert"
-
-	successCertificateIssued  = "CertIssued"
-	successCertificateRenewed = "CertRenewed"
-
-	messageErrorSavingCertificate = "Error saving TLS certificate: "
-)
-
-var (
-	certificateGvk = v1alpha1.SchemeGroupVersion.WithKind("Certificate")
-)
-
-func (c *controller) Sync(ctx context.Context, crt *v1alpha1.Certificate) (err error) {
-	c.metrics.IncrementSyncCallCount(ControllerName)
-
+func (c *certificateRequestManager) ProcessItem(ctx context.Context, key string) error {
 	log := logf.FromContext(ctx)
-	dbg := log.V(logf.DebugLevel)
 
-	if utilfeature.DefaultFeatureGate.Enabled(feature.DisableDeprecatedACMECertificates) &&
-		crt.Spec.ACME != nil {
-		c.recorder.Eventf(crt, corev1.EventTypeWarning, "DeprecatedField", "Deprecated spec.acme field specified and deprecated field feature gate is enabled.")
+	crt, err := getCertificateForKey(ctx, key, c.certificateLister)
+	if apierrors.IsNotFound(err) {
+		log.Error(err, "certificate resource not found for key", "key", key)
 		return nil
 	}
-
-	// if group name is set, use the new experimental controller implementation
-	if crt.Spec.IssuerRef.Group != "" {
-		log.Info("certificate issuerRef group is non-empty, skipping processing")
-		return nil
-	}
-
-	crtCopy := crt.DeepCopy()
-	defer func() {
-		if _, saveErr := updateCertificateStatus(ctx, c.metrics, c.cmClient, crt, crtCopy); saveErr != nil {
-			err = utilerrors.NewAggregate([]error{saveErr, err})
-		}
-	}()
-
-	dbg.Info("Fetching existing certificate from secret", "name", crtCopy.Spec.SecretName)
-	// grab existing certificate and validate private key
-	certs, key, err := kube.SecretTLSKeyPair(ctx, c.secretLister, crtCopy.Namespace, crtCopy.Spec.SecretName)
-	// if we don't have a certificate, we need to trigger a re-issue immediately
-	if err != nil && !(k8sErrors.IsNotFound(err) || errors.IsInvalidData(err)) {
-		return err
-	}
-
-	var cert *x509.Certificate
-	if len(certs) > 0 {
-		dbg.Info("Found existing certificate in secret")
-		cert = certs[0]
-	}
-
-	// update certificate expiry metric
-	defer c.metrics.UpdateCertificateExpiry(crtCopy, c.secretLister)
-	dbg.Info("Update certificate status if required")
-	c.setCertificateStatus(crtCopy, key, cert)
-
-	el := validation.ValidateCertificate(crtCopy)
-	if len(el) > 0 {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, "BadConfig", "Resource validation failed: %v", el.ToAggregate())
-		return nil
-	}
-
-	// check that certificate secret name is unique
-	namespaceCrts := c.certificateLister.Certificates(crtCopy.Namespace)
-	otherCrts, err := namespaceCrts.List(labels.Everything())
-	if err != nil {
-		return err
-	}
-	var duplicate *v1alpha1.Certificate
-	for _, otherCrt := range otherCrts {
-		if otherCrt.Name != crtCopy.Name && otherCrt.Spec.SecretName == crtCopy.Spec.SecretName {
-			duplicate = otherCrt
-			break
-		}
-	}
-	if duplicate != nil {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, errorDuplicateSecretName, "Another Certificate %v already specifies spec.secretName %v, please update the secretName on either Certificate", duplicate.Name, crtCopy.Spec.SecretName)
-		key, err := cache.MetaNamespaceKeyFunc(crtCopy)
-		if err != nil {
-			c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, "KeyError", "Failed to create a key for the Certificate: %v", err)
-			return nil
-		}
-		c.scheduledWorkQueue.Forget(key)
-		apiutil.SetCertificateCondition(crtCopy, v1alpha1.CertificateConditionReady, v1alpha1.ConditionFalse, "DuplicateSecretName", "Another Certificate is using the same secretName")
-		return nil
-	}
-
-	// step zero: check if the referenced issuer exists and is ready
-	issuerObj, err := c.helper.GetGenericIssuer(crtCopy.Spec.IssuerRef, crtCopy.Namespace)
-	if k8sErrors.IsNotFound(err) {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, errorIssuerNotFound, err.Error())
+	if crt == nil {
+		log.Info("certificate resource not found for key", "key", key)
 		return nil
 	}
 	if err != nil {
 		return err
 	}
-	dbg.Info("Fetched issuer resource referenced by certificate", "issuer_name", crtCopy.Spec.IssuerRef.Name)
 
-	el = validation.ValidateCertificateForIssuer(crtCopy, issuerObj)
-	if len(el) > 0 {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, "BadConfig", "Resource validation failed: %v", el.ToAggregate())
-		return nil
-	}
+	log = logf.WithResource(log, crt)
 
-	dbg.Info("Certificate passed all validation checks")
-
-	issuerReady := apiutil.IssuerHasCondition(issuerObj, v1alpha1.IssuerCondition{
-		Type:   v1alpha1.IssuerConditionReady,
-		Status: v1alpha1.ConditionTrue,
-	})
-	if !issuerReady {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, errorIssuerNotReady, "Issuer %s not ready", issuerObj.GetObjectMeta().Name)
-		return nil
-	}
-
-	i, err := c.issuerFactory.IssuerFor(issuerObj)
+	ctx = logf.NewContext(ctx, log)
+	updatedCert := crt.DeepCopy()
+	err = c.processCertificate(ctx, updatedCert)
+	log.V(logf.DebugLevel).Info("check if certificate status update is required")
+	updateStatusErr := c.updateCertificateStatus(ctx, crt, updatedCert)
+	// TODO: combine errors
 	if err != nil {
-		c.recorder.Eventf(crtCopy, corev1.EventTypeWarning, errorIssuerInit, "Internal error initialising issuer: %v", err)
-		return nil
+		return err
 	}
-
-	if isTemporaryCertificate(cert) {
-		dbg.Info("Temporary certificate found - calling 'issue'")
-		return c.issue(ctx, i, crtCopy)
+	if updateStatusErr != nil {
+		return err
 	}
-
-	if key == nil || cert == nil {
-		dbg.Info("Invoking issue function as existing certificate does not exist")
-		return c.issue(ctx, i, crtCopy)
-	}
-
-	// begin checking if the TLS certificate is valid/needs a re-issue or renew
-	matches, matchErrs := certificateMatchesSpec(crtCopy, key, cert, c.secretLister)
-	if !matches {
-		dbg.Info("invoking issue function due to certificate not matching spec", "diff", strings.Join(matchErrs, ", "))
-		return c.issue(ctx, i, crtCopy)
-	}
-
-	// check if the certificate needs renewal
-	needsRenew := c.certificateNeedsRenew(ctx, cert, crt)
-	if needsRenew {
-		dbg.Info("invoking issue function due to certificate needing renewal")
-		return c.issue(ctx, i, crtCopy)
-	}
-	// end checking if the TLS certificate is valid/needs a re-issue or renew
-
-	dbg.Info("Certificate does not need updating. Scheduling renewal.")
-	// If the Certificate is valid and up to date, we schedule a renewal in
-	// the future.
-	scheduleRenewal(ctx, c.secretLister, c.calculateDurationUntilRenew, c.scheduledWorkQueue.Add, crt)
 
 	return nil
 }
 
-// setCertificateStatus will update the status subresource of the certificate.
-// It will not actually submit the resource to the apiserver.
-func (c *controller) setCertificateStatus(crt *v1alpha1.Certificate, key crypto.Signer, cert *x509.Certificate) {
-	if key == nil || cert == nil {
-		apiutil.SetCertificateCondition(crt, v1alpha1.CertificateConditionReady, v1alpha1.ConditionFalse, "NotFound", "Certificate does not exist")
-		return
+func (c *certificateRequestManager) updateCertificateStatus(ctx context.Context, old, crt *cmapi.Certificate) error {
+	log := logf.FromContext(ctx)
+	secretExists := true
+	certs, key, err := kube.SecretTLSKeyPair(ctx, c.secretLister, crt.Namespace, crt.Spec.SecretName)
+	if err != nil {
+		if !apierrors.IsNotFound(err) && !errors.IsInvalidData(err) {
+			return err
+		}
+
+		if apierrors.IsNotFound(err) {
+			secretExists = false
+		}
+	}
+	reqs, err := findCertificateRequestsForCertificate(log, crt, c.certificateRequestLister)
+	if err != nil {
+		return err
+	}
+	var req *cmapi.CertificateRequest
+	if len(reqs) == 1 {
+		req = reqs[0]
+	}
+	var cert *x509.Certificate
+	var certExpired bool
+	if len(certs) > 0 {
+		cert = certs[0]
+		certExpired = cert.NotAfter.Before(c.clock.Now())
 	}
 
-	metaNotAfter := metav1.NewTime(cert.NotAfter)
-	crt.Status.NotAfter = &metaNotAfter
+	var matches bool
+	var matchErrs []string
+	if key != nil && cert != nil {
+		matches, matchErrs = certificateMatchesSpec(crt, key, cert, c.secretLister)
+	}
+
+	isTempCert := isTemporaryCertificate(cert)
+
+	// begin setting certificate status fields
+	if !matches || isTempCert {
+		crt.Status.NotAfter = nil
+	} else {
+		metaNotAfter := metav1.NewTime(cert.NotAfter)
+		crt.Status.NotAfter = &metaNotAfter
+	}
 
 	// Derive & set 'Ready' condition on Certificate resource
-	matches, matchErrs := certificateMatchesSpec(crt, key, cert, c.secretLister)
-	ready := v1alpha1.ConditionFalse
+	ready := cmapi.ConditionFalse
 	reason := ""
 	message := ""
 	switch {
-	case isTemporaryCertificate(cert):
-		reason = "TemporaryCertificate"
-		message = "Certificate issuance in progress. Temporary certificate issued."
-		// clear the NotAfter field as it is not relevant to the user
-		crt.Status.NotAfter = nil
-	case cert.NotAfter.Before(c.clock.Now()):
-		reason = "Expired"
-		message = fmt.Sprintf("Certificate has expired on %s", cert.NotAfter.Format(time.RFC822))
+	case !secretExists || key == nil:
+		reason = "NotFound"
+		message = "Certificate does not exist"
+	case matches && !isTempCert && !certExpired:
+		ready = cmapi.ConditionTrue
+		reason = "Ready"
+		message = "Certificate is up to date and has not expired"
+	case req != nil:
+		reason = "InProgress"
+		message = fmt.Sprintf("Waiting for CertificateRequest %q to complete", req.Name)
+	case cert == nil:
+		reason = "Pending"
+		message = "Certificate pending issuance"
 	case !matches:
 		reason = "DoesNotMatch"
 		message = strings.Join(matchErrs, ", ")
+	case certExpired:
+		reason = "Expired"
+		message = fmt.Sprintf("Certificate has expired on %s", cert.NotAfter.Format(time.RFC822))
+	case isTempCert:
+		reason = "TemporaryCertificate"
+		message = "Certificate issuance in progress. Temporary certificate issued."
 	default:
-		ready = v1alpha1.ConditionTrue
-		reason = "Ready"
-		message = "Certificate is up to date and has not expired"
+		// theoretically, it should not be possible to reach this state.
+		// practically, we may have missed some edge cases above.
+		// print a dump of the current state as a log message so that users can
+		// discover, share and attempt to resolve bugs in this area of code easily.
+		log.Info("unknown certificate state",
+			"secret_exists", secretExists,
+			"matches", matches,
+			"is_temp_cert", isTempCert,
+			"cert_expired", certExpired,
+			"key_is_nil", key == nil,
+			"req_is_nil", req == nil,
+			"cert_is_nil", cert == nil,
+		)
+		ready = cmapi.ConditionFalse
+		reason = "Unknown"
+		message = "Unknown certificate status. Please open an issue and share your controller logs."
 	}
+	apiutil.SetCertificateCondition(crt, cmapi.CertificateConditionReady, ready, reason, message)
 
-	apiutil.SetCertificateCondition(crt, v1alpha1.CertificateConditionReady, ready, reason, message)
-
-	return
-}
-
-func certificateMatchesSpec(crt *v1alpha1.Certificate, key crypto.Signer, cert *x509.Certificate, secretLister corelisters.SecretLister) (bool, []string) {
-	var errs []string
-
-	// TODO: add checks for KeySize, KeyAlgorithm fields
-	// TODO: add checks for Organization field
-	// TODO: add checks for IsCA field
-
-	// check if the private key is the corresponding pair to the certificate
-	matches, err := pki.PublicKeyMatchesCertificate(key.Public(), cert)
+	_, err = updateCertificateStatus(ctx, metrics.Default, c.cmClient, old, crt)
 	if err != nil {
-		errs = append(errs, err.Error())
-	} else if !matches {
-		errs = append(errs, fmt.Sprintf("Certificate private key does not match certificate"))
-	}
-
-	// validate the common name is correct
-	expectedCN := pki.CommonNameForCertificate(crt)
-	if expectedCN != cert.Subject.CommonName {
-		errs = append(errs, fmt.Sprintf("Common name on TLS certificate not up to date: %q", cert.Subject.CommonName))
-	}
-
-	// validate the dns names are correct
-	expectedDNSNames := pki.DNSNamesForCertificate(crt)
-	if !util.EqualUnsorted(cert.DNSNames, expectedDNSNames) {
-		errs = append(errs, fmt.Sprintf("DNS names on TLS certificate not up to date: %q", cert.DNSNames))
-	}
-
-	// validate the ip addresses are correct
-	if !util.EqualUnsorted(pki.IPAddressesToString(cert.IPAddresses), crt.Spec.IPAddresses) {
-		errs = append(errs, fmt.Sprintf("IP addresses on TLS certificate not up to date: %q", pki.IPAddressesToString(cert.IPAddresses)))
-	}
-
-	// get a copy of the current secret resource
-	// Note that we already know that it exists, no need to check for errors
-	// TODO: Refactor so that the secret is passed as argument?
-	secret, err := secretLister.Secrets(crt.Namespace).Get(crt.Spec.SecretName)
-
-	// validate that the issuer is correct
-	if crt.Spec.IssuerRef.Name != secret.Annotations[v1alpha1.IssuerNameAnnotationKey] {
-		errs = append(errs, fmt.Sprintf("Issuer of the certificate is not up to date: %q", secret.Annotations[v1alpha1.IssuerNameAnnotationKey]))
-	}
-
-	// validate that the issuer kind is correct
-	if apiutil.IssuerKind(crt.Spec.IssuerRef) != secret.Annotations[v1alpha1.IssuerKindAnnotationKey] {
-		errs = append(errs, fmt.Sprintf("Issuer kind of the certificate is not up to date: %q", secret.Annotations[v1alpha1.IssuerKindAnnotationKey]))
-	}
-
-	return len(errs) == 0, errs
-}
-
-func scheduleRenewal(ctx context.Context, lister corelisters.SecretLister, calc calculateDurationUntilRenewFn, queueFn func(interface{}, time.Duration), crt *v1alpha1.Certificate) {
-	log := logf.FromContext(ctx)
-	log = log.WithValues(
-		logf.RelatedResourceNameKey, crt.Spec.SecretName,
-		logf.RelatedResourceNamespaceKey, crt.Namespace,
-		logf.RelatedResourceKindKey, "Secret",
-	)
-
-	key, err := keyFunc(crt)
-	if err != nil {
-		log.Error(err, "error getting key for certificate resource")
-		return
-	}
-
-	cert, err := kube.SecretTLSCert(ctx, lister, crt.Namespace, crt.Spec.SecretName)
-	if err != nil {
-		if !errors.IsInvalidData(err) {
-			log.Error(err, "error getting secret for certificate resource")
-		}
-		return
-	}
-
-	renewIn := calc(ctx, cert, crt)
-	queueFn(key, renewIn)
-
-	log.WithValues("duration_until_renewal", renewIn.String()).Info("certificate scheduled for renewal")
-}
-
-func ownerRef(crt *v1alpha1.Certificate) metav1.OwnerReference {
-	controller := true
-	return metav1.OwnerReference{
-		APIVersion: v1alpha1.SchemeGroupVersion.String(),
-		Kind:       v1alpha1.CertificateKind,
-		Name:       crt.Name,
-		UID:        crt.UID,
-		Controller: &controller,
-	}
-}
-
-// updateSecret will store the provided secret data into the target secret
-// named on the Certificate resource.
-// - If the secret is empty, a new one will be created containing the data
-// - If a secret already exists, its contents will be overwritten
-// - If the provided certificate is a temporary certificate and the certificate
-//   stored in the secret is already a temporary certificate, then the Secret
-//   **will not** be updated.
-func (c *controller) updateSecret(ctx context.Context, crt *v1alpha1.Certificate, namespace string, cert, key, ca []byte) (*corev1.Secret, error) {
-	log := logf.FromContext(ctx, "updateSecret")
-	log = logf.WithRelatedResourceName(log, crt.Spec.SecretName, namespace, "Secret")
-
-	// if the key is not set, we bail out early.
-	// this function should always be called with at least a private key.
-	// in future we'll likely need to relax this requirement, but for now we'll
-	// keep this here to be safe.
-	if len(key) == 0 {
-		return nil, fmt.Errorf("private key data must be set")
-	}
-	privKey, err := pki.DecodePrivateKeyBytes(key)
-	if err != nil {
-		return nil, fmt.Errorf("error decoding private key: %v", err)
-	}
-
-	// get a copy of the current secret resource
-	secret, err := c.secretLister.Secrets(namespace).Get(crt.Spec.SecretName)
-	if err != nil && !k8sErrors.IsNotFound(err) {
-		return nil, err
-	}
-	// create a deep copy of the secret before modifying it as we fetched it
-	// from the lister's cache.
-	if secret != nil {
-		secret = secret.DeepCopy()
-	}
-	// if the resource does not already exist, we will create a new one
-	if secret == nil {
-		secret = &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      crt.Spec.SecretName,
-				Namespace: namespace,
-			},
-			Type: corev1.SecretTypeTLS,
-			Data: map[string][]byte{},
-		}
-	}
-	// include this clause in case the existing secret has nil data
-	if secret.Data == nil {
-		secret.Data = map[string][]byte{}
-	}
-	if secret.Annotations == nil {
-		secret.Annotations = make(map[string]string)
-	}
-	if secret.Labels == nil {
-		secret.Labels = make(map[string]string)
-	}
-
-	var existingCert *x509.Certificate
-	existingCertData := secret.Data[corev1.TLSCertKey]
-	if len(existingCertData) > 0 {
-		existingCert, err = pki.DecodeX509CertificateBytes(existingCertData)
-		if err != nil {
-			log.Error(err, "error decoding existing x509 certificate bytes, continuing anyway")
-		}
-	}
-
-	var x509Cert *x509.Certificate
-	switch {
-	case len(cert) > 0:
-		x509Cert, err = pki.DecodeX509CertificateBytes(cert)
-		if err != nil {
-			return nil, fmt.Errorf("invalid certificate data: %v", err)
-		}
-	case !utilfeature.DefaultFeatureGate.Enabled(feature.IssueTemporaryCertificate):
-		break
-	case isTemporaryCertificate(existingCert):
-		matches, err := pki.PublicKeyMatchesCertificate(privKey.Public(), existingCert)
-		if err == nil && matches {
-			// if the existing certificate is a temporary one, and the certificate
-			// being written in this call to updateSecret is not set, then we do
-			// not want to keep re-issuing new temporary certificates
-			cert = existingCertData
-			x509Cert = existingCert
-			break
-		}
-		fallthrough
-	default:
-		// if the issuer returns a private key but not certificate data, we
-		// generate and store a temporary certificate that we can later
-		// recognise and force later calls to the issuer's Issue method
-		cert, err = c.localTemporarySigner(crt, key)
-		if err != nil {
-			return nil, fmt.Errorf("error signing locally generated certificate: %v", err)
-		}
-
-		x509Cert, err = pki.DecodeX509CertificateBytes(cert)
-		if err != nil {
-			return nil, fmt.Errorf("invalid certificate data: %v", err)
-		}
-
-		c.recorder.Event(crt, corev1.EventTypeNormal, "GenerateSelfSigned", "Generated temporary self signed certificate")
-	}
-
-	// TODO: move metadata setting out of this method, and support
-	// retrospectively adding metadata annotations on every Sync iteration and
-	// not just when a new certificate is issued
-	if x509Cert != nil {
-		secret.Annotations[v1alpha1.IssuerNameAnnotationKey] = crt.Spec.IssuerRef.Name
-		secret.Annotations[v1alpha1.IssuerKindAnnotationKey] = apiutil.IssuerKind(crt.Spec.IssuerRef)
-		secret.Annotations[v1alpha1.CommonNameAnnotationKey] = x509Cert.Subject.CommonName
-		secret.Annotations[v1alpha1.AltNamesAnnotationKey] = strings.Join(x509Cert.DNSNames, ",")
-		secret.Annotations[v1alpha1.IPSANAnnotationKey] = strings.Join(pki.IPAddressesToString(x509Cert.IPAddresses), ",")
-		secret.Annotations[v1alpha1.CertificateNameKey] = crt.Name
-	}
-
-	// Always set the certificate name label on the target secret
-	// TODO: remove this behaviour - there is a max length limit of 64 chars on label values which causes issues here
-	secret.Labels[v1alpha1.CertificateNameKey] = crt.Name
-
-	// set the actual values in the secret
-	secret.Data[corev1.TLSCertKey] = cert
-	secret.Data[corev1.TLSPrivateKeyKey] = key
-	secret.Data[v1alpha1.TLSCAKey] = ca
-
-	// if it is a new resource
-	if secret.SelfLink == "" {
-		if c.addOwnerReferences {
-			secret.SetOwnerReferences(append(secret.GetOwnerReferences(), ownerRef(crt)))
-		}
-		secret, err = c.kClient.CoreV1().Secrets(namespace).Create(secret)
-	} else {
-		secret, err = c.kClient.CoreV1().Secrets(namespace).Update(secret)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return secret, nil
-}
-
-// return an error on failure. If retrieval is succesful, the certificate data
-// and private key will be stored in the named secret
-func (c *controller) issue(ctx context.Context, issuer issuer.Interface, crt *v1alpha1.Certificate) error {
-	log := logf.FromContext(ctx)
-
-	resp, err := issuer.Issue(ctx, crt)
-	if err != nil {
-		log.Error(err, "error issuing certificate")
 		return err
-	}
-	// if the issuer has not returned any data, exit early
-	if resp == nil {
-		return nil
-	}
-
-	if _, err := c.updateSecret(ctx, crt, crt.Namespace, resp.Certificate, resp.PrivateKey, resp.CA); err != nil {
-		s := messageErrorSavingCertificate + err.Error()
-		log.Error(err, "error saving certificate")
-		c.recorder.Event(crt, corev1.EventTypeWarning, errorSavingCertificate, s)
-		return err
-	}
-
-	if len(resp.Certificate) > 0 {
-		c.recorder.Event(crt, corev1.EventTypeNormal, successCertificateIssued, "Certificate issued successfully")
-		// as we have just written a certificate, we should schedule it for renewal
-		scheduleRenewal(ctx, c.secretLister, c.calculateDurationUntilRenew, c.scheduledWorkQueue.Add, crt)
 	}
 
 	return nil
 }
 
-// staticTemporarySerialNumber is a fixed serial number we check for when
-// updating the status of a certificate.
-// It is used to identify temporarily generated certificates, so that friendly
-// status messages can be displayed to users.
-const staticTemporarySerialNumber = 0x1234567890
+// processCertificate is the core method that is called in the manager.
+// It accepts a Certificate resource, and checks to see if the certificate
+// requires re-issuance.
+func (c *certificateRequestManager) processCertificate(ctx context.Context, crt *cmapi.Certificate) error {
+	log := logf.FromContext(ctx)
+	dbg := log.V(logf.DebugLevel)
 
-func isTemporaryCertificate(cert *x509.Certificate) bool {
-	if cert == nil {
-		return false
+	// The certificate request name is a product of the certificate's spec,
+	// which makes it unique and predictable.
+	// First we compute what we expect it to be.
+	expectedReqName, err := expectedCertificateRequestName(crt)
+	if err != nil {
+		return fmt.Errorf("internal error hashing certificate spec: %v", err)
 	}
-	return cert.SerialNumber.Int64() == staticTemporarySerialNumber
+
+	// Clean up any 'owned' CertificateRequest resources that do not have the
+	// expected name computed above
+	err = c.cleanupExistingCertificateRequests(log, crt, expectedReqName)
+	if err != nil {
+		return err
+	}
+
+	// Fetch a copy of the existing Secret resource
+	existingSecret, err := c.secretLister.Secrets(crt.Namespace).Get(crt.Spec.SecretName)
+	if apierrors.IsNotFound(err) {
+		// If the secret does not exist, generate a new private key and store it.
+		dbg.Info("existing secret not found, generating and storing private key")
+		return c.generateAndStorePrivateKey(ctx, crt, nil, c.kubeClient.CoreV1().Secrets(crt.Namespace).Create)
+	}
+	if err != nil {
+		return err
+	}
+
+	log = logf.WithRelatedResource(log, existingSecret)
+	ctx = logf.NewContext(ctx, log)
+
+	// If the Secret does not contain a private key, generate one and update
+	// the Secret resource
+	existingKey := existingSecret.Data[corev1.TLSPrivateKeyKey]
+	if len(existingKey) == 0 {
+		log.Info("existing private key not found in Secret, generate a new private key")
+		return c.generateAndStorePrivateKey(ctx, crt, existingSecret, c.kubeClient.CoreV1().Secrets(crt.Namespace).Update)
+	}
+
+	// Ensure the the private key has the correct key algorithm and key size.
+	dbg.Info("validating private key has correct keyAlgorithm/keySize")
+	validKey, err := validatePrivateKeyUpToDate(log, existingKey, crt)
+	// If tls.key contains invalid data, we regenerate a new private key
+	if errors.IsInvalidData(err) {
+		log.Info("existing private key data is invalid, generating a new private key")
+		return c.generateAndStorePrivateKey(ctx, crt, existingSecret, c.kubeClient.CoreV1().Secrets(crt.Namespace).Update)
+	}
+	if err != nil {
+		return err
+	}
+	// If the private key is not 'up to date', we generate a new private key
+	if !validKey {
+		log.Info("existing private key does not match requirements specified on Certificate resource, generating new private key")
+		return c.generateAndStorePrivateKey(ctx, crt, existingSecret, c.kubeClient.CoreV1().Secrets(crt.Namespace).Update)
+	}
+
+	// Attempt to fetch the CertificateRequest with the expected name computed above.
+	dbg.Info("checking for existing CertificateRequest for Certificate")
+	existingReq, err := c.certificateRequestLister.CertificateRequests(crt.Namespace).Get(expectedReqName)
+	// Allow IsNotFound errors, later on we check if existingReq == nil and if
+	// it is, we create a new CertificateRequest resource.
+	if err != nil && !apierrors.IsNotFound(err) {
+		return err
+	}
+	if existingReq != nil {
+		dbg.Info("found existing certificate request for Certificate", "request_name", existingReq.Name)
+		log = logf.WithRelatedResource(log, existingReq)
+	}
+
+	needsIssue := true
+	// Parse the existing certificate
+	existingCert := existingSecret.Data[corev1.TLSCertKey]
+	if len(existingCert) > 0 {
+		// Here we check to see if the existing certificate 'matches' the spec
+		// of the Certificate resource.
+		// This includes checking if dnsNames, commonName, organization etc.
+		// are up to date, as well as validating that the stored private key is
+		// a valid partner to the stored certificate.
+		var matchErrs []string
+		dbg.Info("checking if existing certificate stored in Secret resource is not expiring soon and matches certificate spec")
+		needsIssue, matchErrs, err = c.certificateRequiresIssuance(ctx, crt, existingKey, existingCert)
+		if err != nil && !errors.IsInvalidData(err) {
+			return err
+		}
+		// If the certificate data is invalid, we require a re-issuance.
+		// The private key should never be invalid at this point as we already
+		// check it above.
+		if errors.IsInvalidData(err) {
+			dbg.Info("existing secret contains invalid certificate data")
+			needsIssue = true
+		}
+
+		if !needsIssue {
+			dbg.Info("existing certificate does not need re-issuance")
+		} else {
+			dbg.Info("will attempt to issue certificate", "reason", matchErrs)
+		}
+	}
+
+	// Exit early if the certificate doesn't need issuing to save extra work
+	if !needsIssue {
+		if existingReq != nil {
+			dbg.Info("skipping issuing certificate data into Secret resource as existing issued certificate is still valid")
+		}
+
+		// Before exiting, ensure that the Secret resource's metadata is up to
+		// date. If it isn't, it will be updated.
+		updated, err := c.ensureSecretMetadataUpToDate(ctx, existingSecret, crt)
+		if err != nil {
+			return err
+		}
+
+		if updated {
+			log.Info("updated Secret resource metadata as it was out of date")
+		}
+
+		// As the Certificate has been validated as Ready, schedule a renewal
+		// for near the expiry date.
+		scheduleRenewal(ctx, c.secretLister, c.calculateDurationUntilRenew, c.scheduledWorkQueue.Add, crt)
+
+		log.Info("certificate does not require re-issuance. certificate renewal scheduled near expiry time.")
+
+		return nil
+	}
+
+	// Attempt to decode the private key.
+	// This shouldn't fail as we already validate the private key is valid above.
+	dbg.Info("decoding existing private key")
+	privateKey, err := pki.DecodePrivateKeyBytes(existingKey)
+	if err != nil {
+		return err
+	}
+
+	// Attempt to decode the existing certificate.
+	// We tolerate invalid data errors as we will issue a certificate if the
+	// data is invalid.
+	dbg.Info("attempting to decode existing certificate")
+	existingX509Cert, err := pki.DecodeX509CertificateBytes(existingCert)
+	if err != nil && !errors.IsInvalidData(err) {
+		return err
+	}
+	if errors.IsInvalidData(err) {
+		dbg.Info("existing certificate data is invalid, continuing...")
+	}
+
+	// Handling for 'temporary certificates'
+	if c.issueTemporaryCerts {
+		// Issue a temporary certificate if the current certificate is empty or the
+		// private key is not valid for the current certificate.
+		if existingX509Cert == nil {
+			log.Info("no existing certificate data found in secret, issuing temporary certificate")
+			return c.issueTemporaryCertificate(ctx, existingSecret, crt, existingKey)
+		}
+		// We don't issue a temporary certificate if the existing stored
+		// certificate already 'matches', even if it isn't a temporary certificate.
+		matches, _ := certificateMatchesSpec(crt, privateKey, existingX509Cert, c.secretLister)
+		if !matches {
+			log.Info("existing certificate fields do not match certificate spec, issuing temporary certificate")
+			return c.issueTemporaryCertificate(ctx, existingSecret, crt, existingKey)
+		}
+
+		log.Info("not issuing temporary certificate as existing certificate matches requirements")
+
+		// Ensure the secret metadata is up to date
+		updated, err := c.ensureSecretMetadataUpToDate(ctx, existingSecret, crt)
+		if err != nil {
+			return err
+		}
+
+		// Only return early if an update actually occurred, otherwise continue.
+		if updated {
+			log.Info("updated Secret resource metadata as it was out of date")
+			return nil
+		}
+	}
+
+	if existingReq == nil {
+		// If no existing CertificateRequest resource exists, we must create one
+		log.Info("no existing CertificateRequest resource exists, creating new request...")
+		req, err := c.buildCertificateRequest(log, crt, expectedReqName, existingKey)
+		if err != nil {
+			return err
+		}
+
+		req, err = c.cmClient.CertmanagerV1alpha1().CertificateRequests(crt.Namespace).Create(req)
+		if err != nil {
+			return err
+		}
+
+		c.recorder.Eventf(crt, corev1.EventTypeNormal, "Requested", "Created new CertificateRequest resource %q", req.Name)
+		log.Info("created certificate request", "request_name", req.Name)
+
+		return nil
+	}
+
+	// Validate the CertificateRequest's CSR is valid
+	log.Info("validating existing CSR data")
+	x509CSR, err := pki.DecodeX509CertificateRequestBytes(existingReq.Spec.CSRPEM)
+	// TODO: handle InvalidData
+	if err != nil {
+		return err
+	}
+
+	// Ensure the stored private key is a 'pair' to the CSR
+	publicKeyMatches, err := pki.PublicKeyMatchesCSR(privateKey.Public(), x509CSR)
+	if err != nil {
+		return err
+	}
+
+	// if the stored private key does not pair with the CSR on the
+	// CertificateRequest resource, delete the resource as we won't be able to
+	// do anything with the certificate if it is issued
+	if !publicKeyMatches {
+		log.Info("stored private key is not valid for CSR stored on existing CertificateRequest, recreating CertificateRequest resource")
+		err := c.cmClient.CertmanagerV1alpha1().CertificateRequests(existingReq.Namespace).Delete(existingReq.Name, nil)
+		if err != nil {
+			return err
+		}
+
+		c.recorder.Eventf(crt, corev1.EventTypeNormal, "PrivateKeyLost", "Lost private key for CertificateRequest %q, deleting old resource", existingReq.Name)
+		log.Info("deleted existing CertificateRequest as the stored private key does not match the CSR")
+		return nil
+	}
+
+	reason := apiutil.CertificateRequestReadyReason(existingReq)
+	// Determine the status reason of the CertificateRequest and process accordingly
+	switch reason {
+	// If the CertificateRequest exists but has failed then we check the failure
+	// time. If the failure time doesn't exist or is over an hour in the past
+	// then delete the request so it can be re-created on the next sync. If the
+	// failure time is less than an hour in the past then schedule this owning
+	// Certificate for a re-sync in an hour.
+	case cmapi.CertificateRequestReasonFailed:
+		if existingReq.Status.FailureTime == nil || c.clock.Since(existingReq.Status.FailureTime.Time) > time.Hour {
+			log.Info("deleting failed certificate request")
+			err := c.cmClient.CertmanagerV1alpha1().CertificateRequests(existingReq.Namespace).Delete(existingReq.Name, nil)
+			if err != nil {
+				return err
+			}
+
+			c.recorder.Eventf(crt, corev1.EventTypeNormal, "CertificateRequestRetry", "The failed CertificateRequest %q will be retried now", existingReq.Name)
+			return nil
+		}
+
+		log.Info("the failed existing certificate request failed less than an hour ago, will be scheduled for reprocessing in an hour")
+
+		key, err := keyFunc(crt)
+		if err != nil {
+			log.Error(err, "error getting key for certificate resource")
+			return nil
+		}
+
+		// We don't fire an event here as this could be called multiple times in quick succession
+		c.scheduledWorkQueue.Add(key, time.Hour)
+		return nil
+
+		// If the CertificateRequest is in a Ready state then we can decode,
+		// verify, and check whether it needs renewal
+	case cmapi.CertificateRequestReasonIssued:
+		log.Info("CertificateRequest is in a Ready state, issuing certificate...")
+
+		// Decode the certificate bytes so we can ensure the certificate is valid
+		log.Info("decoding certificate data")
+		x509Cert, err := pki.DecodeX509CertificateBytes(existingReq.Status.Certificate)
+		if err != nil {
+			return err
+		}
+
+		// Check if the Certificate requires renewal according to the renewBefore
+		// specified on the Certificate resource.
+		log.Info("checking if certificate stored on CertificateRequest is up to date")
+		if c.certificateNeedsRenew(ctx, x509Cert, crt) {
+			log.Info("certificate stored on CertificateRequest needs renewal, so deleting the old CertificateRequest resource")
+			err := c.cmClient.CertmanagerV1alpha1().CertificateRequests(existingReq.Namespace).Delete(existingReq.Name, nil)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		}
+
+		// If certificate stored on CertificateRequest is not expiring soon, copy
+		// across the status.certificate field into the Secret resource.
+		log.Info("CertificateRequest contains a valid certificate for issuance. Issuing certificate...")
+
+		_, err = c.updateSecretData(ctx, crt, existingSecret, secretData{pk: existingKey, cert: existingReq.Status.Certificate, ca: existingReq.Status.CA})
+		if err != nil {
+			return err
+		}
+
+		c.recorder.Eventf(crt, corev1.EventTypeNormal, "Issued", "Certificate issued successfully")
+		return nil
+
+		// If it is not Ready _OR_ Failed then we return and wait for informer
+		// updates to re-trigger processing.
+	default:
+		log.Info("CertificateRequest is not in a final state, waiting until CertificateRequest is complete", "state", reason)
+		return nil
+	}
 }
 
-func generateSelfSignedTemporaryCertificate(crt *v1alpha1.Certificate, pk []byte) ([]byte, error) {
-	template, err := pki.GenerateTemplate(crt)
-	template.SerialNumber = big.NewInt(staticTemporarySerialNumber)
+// updateSecretData will ensure the Secret resource contains the given secret
+// data as well as appropriate metadata.
+// If the given 'existingSecret' is nil, a new Secret resource will be created.
+// Otherwise, the existing resource will be updated.
+// The first return argument will be true if the resource was updated/created
+// without error.
+func (c *certificateRequestManager) updateSecretData(ctx context.Context, crt *cmapi.Certificate, existingSecret *corev1.Secret, data secretData) (bool, error) {
+	s := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      crt.Spec.SecretName,
+			Namespace: crt.Namespace,
+		},
+		Type: corev1.SecretTypeTLS,
+	}
+	if existingSecret != nil {
+		s = existingSecret
+	}
+
+	newSecret := s.DeepCopy()
+	err := setSecretValues(ctx, crt, newSecret, secretData{pk: data.pk, cert: data.cert, ca: data.ca})
+	if err != nil {
+		return false, err
+	}
+	if reflect.DeepEqual(s, newSecret) {
+		return false, nil
+	}
+
+	if existingSecret == nil {
+		_, err = c.kubeClient.CoreV1().Secrets(newSecret.Namespace).Create(newSecret)
+		if err != nil {
+			return false, err
+		}
+		return true, nil
+	}
+
+	_, err = c.kubeClient.CoreV1().Secrets(newSecret.Namespace).Update(newSecret)
+	if err != nil {
+		return false, err
+	}
+
+	return true, nil
+}
+
+func (c *certificateRequestManager) ensureSecretMetadataUpToDate(ctx context.Context, s *corev1.Secret, crt *cmapi.Certificate) (bool, error) {
+	pk := s.Data[corev1.TLSPrivateKeyKey]
+	cert := s.Data[corev1.TLSCertKey]
+	ca := s.Data[cmapi.TLSCAKey]
+
+	updated, err := c.updateSecretData(ctx, crt, s, secretData{pk: pk, cert: cert, ca: ca})
+	if err != nil || !updated {
+		return updated, err
+	}
+
+	c.recorder.Eventf(crt, corev1.EventTypeNormal, "UpdateMeta", "Updated metadata on Secret resource")
+
+	return true, nil
+}
+
+func (c *certificateRequestManager) issueTemporaryCertificate(ctx context.Context, secret *corev1.Secret, crt *cmapi.Certificate, key []byte) error {
+	tempCertData, err := c.localTemporarySigner(crt, key)
+	if err != nil {
+		return err
+	}
+
+	newSecret := secret.DeepCopy()
+	err = setSecretValues(ctx, crt, newSecret, secretData{pk: key, cert: tempCertData})
+	if err != nil {
+		return err
+	}
+
+	newSecret, err = c.kubeClient.CoreV1().Secrets(newSecret.Namespace).Update(newSecret)
+	if err != nil {
+		return err
+	}
+
+	c.recorder.Eventf(crt, corev1.EventTypeNormal, "TempCert", "Issued temporary certificate")
+
+	return nil
+}
+
+func (c *certificateRequestManager) certificateRequiresIssuance(ctx context.Context, crt *cmapi.Certificate, keyBytes, certBytes []byte) (bool, []string, error) {
+	key, err := pki.DecodePrivateKeyBytes(keyBytes)
+	if err != nil {
+		return false, nil, err
+	}
+	cert, err := pki.DecodeX509CertificateBytes(certBytes)
+	if err != nil {
+		return false, nil, err
+	}
+	if isTemporaryCertificate(cert) {
+		return true, nil, nil
+	}
+	matches, matchErrs := certificateMatchesSpec(crt, key, cert, c.secretLister)
+	if !matches {
+		return true, matchErrs, nil
+	}
+	needsRenew := c.certificateNeedsRenew(ctx, cert, crt)
+	return needsRenew, []string{"Certificate is expiring soon"}, nil
+}
+
+func expectedCertificateRequestName(crt *cmapi.Certificate) (string, error) {
+	crt = crt.DeepCopy()
+	specBytes, err := json.Marshal(crt.Spec)
+	if err != nil {
+		return "", err
+	}
+
+	hashF := fnv.New32()
+	_, err = hashF.Write(specBytes)
+	if err != nil {
+		return "", err
+	}
+
+	// shorten the cert name to 52 chars to ensure the total length of the name
+	// is less than or equal to 64 characters
+	return fmt.Sprintf("%.52s-%d", crt.Name, hashF.Sum32()), nil
+}
+
+type generateCSRFn func(*cmapi.Certificate, []byte) ([]byte, error)
+
+func generateCSRImpl(crt *cmapi.Certificate, pk []byte) ([]byte, error) {
+	csr, err := pki.GenerateCSR(crt)
+	if err != nil {
+		return nil, err
+	}
 
 	signer, err := pki.DecodePrivateKeyBytes(pk)
 	if err != nil {
 		return nil, err
 	}
 
-	b, _, err := pki.SignCertificate(template, template, signer.Public(), signer)
+	csrDER, err := pki.EncodeCSR(csr, signer)
 	if err != nil {
 		return nil, err
 	}
 
-	return b, nil
-}
-
-// generateLocallySignedTemporaryCertificate signs a temporary certificate for
-// the given certificate resource using a one-use temporary CA that is then
-// discarded afterwards.
-// This is to mitigate a potential attack against x509 certificates that use a
-// predictable serial number and weak MD5 hashing algorithms.
-// In practice, this shouldn't really be a concern anyway.
-func generateLocallySignedTemporaryCertificate(crt *v1alpha1.Certificate, pk []byte) ([]byte, error) {
-	// generate a throwaway self-signed root CA
-	caPk, err := pki.GenerateECPrivateKey(pki.ECCurve521)
-	if err != nil {
-		return nil, err
-	}
-	caCertTemplate, err := pki.GenerateTemplate(&v1alpha1.Certificate{
-		Spec: v1alpha1.CertificateSpec{
-			CommonName: "cert-manager.local",
-			IsCA:       true,
-		},
+	csrPEM := pem.EncodeToMemory(&pem.Block{
+		Type: "CERTIFICATE REQUEST", Bytes: csrDER,
 	})
-	if err != nil {
-		return nil, err
-	}
-	_, caCert, err := pki.SignCertificate(caCertTemplate, caCertTemplate, caPk.Public(), caPk)
-	if err != nil {
-		return nil, err
-	}
 
-	// sign a temporary certificate using the root CA
-	template, err := pki.GenerateTemplate(crt)
-	if err != nil {
-		return nil, err
-	}
-	template.SerialNumber = big.NewInt(staticTemporarySerialNumber)
-
-	signeeKey, err := pki.DecodePrivateKeyBytes(pk)
-	if err != nil {
-		return nil, err
-	}
-
-	b, _, err := pki.SignCertificate(template, caCert, signeeKey.Public(), caPk)
-	if err != nil {
-		return nil, err
-	}
-
-	return b, nil
+	return csrPEM, nil
 }
 
-func updateCertificateStatus(ctx context.Context, m *metrics.Metrics, cmClient cmclient.Interface, old, new *v1alpha1.Certificate) (*v1alpha1.Certificate, error) {
-	defer m.UpdateCertificateStatus(new)
-
-	log := logf.FromContext(ctx, "updateStatus")
-	oldBytes, _ := json.Marshal(old.Status)
-	newBytes, _ := json.Marshal(new.Status)
-	if reflect.DeepEqual(oldBytes, newBytes) {
-		return nil, nil
+func (c *certificateRequestManager) buildCertificateRequest(log logr.Logger, crt *cmapi.Certificate, name string, pk []byte) (*cmapi.CertificateRequest, error) {
+	csrPEM, err := c.generateCSR(crt, pk)
+	if err != nil {
+		return nil, err
 	}
-	log.V(logf.DebugLevel).Info("updating resource due to change in status", "diff", pretty.Diff(string(oldBytes), string(newBytes)))
-	// TODO: replace Update call with UpdateStatus. This requires a custom API
-	// server with the /status subresource enabled and/or subresource support
-	// for CRDs (https://github.com/kubernetes/kubernetes/issues/38113)
-	return cmClient.CertmanagerV1alpha1().Certificates(new.Namespace).Update(new)
+
+	cr := &cmapi.CertificateRequest{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:            name,
+			Namespace:       crt.Namespace,
+			OwnerReferences: []metav1.OwnerReference{*metav1.NewControllerRef(crt, certificateGvk)},
+			Annotations: map[string]string{
+				cmapi.CRPrivateKeyAnnotationKey: crt.Spec.SecretName,
+			},
+			Labels: crt.Labels,
+		},
+		Spec: cmapi.CertificateRequestSpec{
+			CSRPEM:    csrPEM,
+			Duration:  crt.Spec.Duration,
+			IssuerRef: crt.Spec.IssuerRef,
+			IsCA:      crt.Spec.IsCA,
+		},
+	}
+
+	return cr, nil
+}
+
+func (c *certificateRequestManager) cleanupExistingCertificateRequests(log logr.Logger, crt *cmapi.Certificate, retain string) error {
+	reqs, err := findCertificateRequestsForCertificate(log, crt, c.certificateRequestLister)
+	if err != nil {
+		return err
+	}
+
+	for _, req := range reqs {
+		log := logf.WithRelatedResource(log, req)
+		if req.Name == retain {
+			log.V(logf.DebugLevel).Info("skipping deleting CertificateRequest as it is up to date for the certificate spec")
+			continue
+		}
+
+		err = c.cmClient.CertmanagerV1alpha1().CertificateRequests(req.Namespace).Delete(req.Name, nil)
+		if err != nil {
+			return err
+		}
+
+		log.Info("deleted no longer required CertificateRequest")
+	}
+
+	return nil
+}
+
+func findCertificateRequestsForCertificate(log logr.Logger, crt *cmapi.Certificate, lister cmlisters.CertificateRequestLister) ([]*cmapi.CertificateRequest, error) {
+	log.V(logf.DebugLevel).Info("finding existing CertificateRequest resources for Certificate")
+	reqs, err := lister.CertificateRequests(crt.Namespace).List(labels.Everything())
+	if err != nil {
+		return nil, err
+	}
+
+	var candidates []*cmapi.CertificateRequest
+	for _, req := range reqs {
+		log := logf.WithRelatedResource(log, req)
+		if metav1.IsControlledBy(req, crt) {
+			log.V(logf.DebugLevel).Info("found CertificateRequest resource for Certificate")
+			candidates = append(candidates, &(*req))
+		}
+	}
+
+	return candidates, nil
+}
+
+// validatePrivateKeyUpToDate will evaluate the private key data in pk and
+// ensure it is 'up to date' and matches the specification of the key as
+// required by the given Certificate resource.
+// It returns false if the private key isn't up to date, e.g. the Certificate
+// resource specifies a different keyEncoding, keyAlgorithm or keySize.
+func validatePrivateKeyUpToDate(log logr.Logger, pk []byte, crt *cmapi.Certificate) (bool, error) {
+	signer, err := pki.DecodePrivateKeyBytes(pk)
+	if err != nil {
+		return false, err
+	}
+
+	// TODO: check keyEncoding
+
+	wantedAlgorithm := crt.Spec.KeyAlgorithm
+	if wantedAlgorithm == "" {
+		// in-memory defaulting of the key algorithm to RSA
+		// TODO: remove this in favour of actual defaulting in a mutating webhook
+		wantedAlgorithm = cmapi.RSAKeyAlgorithm
+	}
+
+	switch wantedAlgorithm {
+	case cmapi.RSAKeyAlgorithm:
+		_, ok := signer.(*rsa.PrivateKey)
+		if !ok {
+			log.Info("expected private key's algorithm to be RSA but it is not")
+			return false, nil
+		}
+	// TODO: check keySize
+	case cmapi.ECDSAKeyAlgorithm:
+		_, ok := signer.(*ecdsa.PrivateKey)
+		if !ok {
+			log.Info("expected private key's algorithm to be ECDSA but it is not")
+			return false, nil
+		}
+		// TODO: check keySize
+	}
+
+	return true, nil
+}
+
+type secretSaveFn func(*corev1.Secret) (*corev1.Secret, error)
+
+func (c *certificateRequestManager) generateAndStorePrivateKey(ctx context.Context, crt *cmapi.Certificate, s *corev1.Secret, saveFn secretSaveFn) error {
+	keyData, err := c.generatePrivateKeyBytes(ctx, crt)
+	if err != nil {
+		// TODO: handle permanent failures caused by invalid spec
+		return err
+	}
+
+	updated, err := c.updateSecretData(ctx, crt, s, secretData{pk: keyData})
+	if err != nil {
+		return err
+	}
+	if !updated {
+		return nil
+	}
+
+	c.recorder.Eventf(crt, corev1.EventTypeNormal, "GeneratedKey", "Generated a new private key")
+
+	return nil
+}
+
+type generatePrivateKeyBytesFn func(context.Context, *cmapi.Certificate) ([]byte, error)
+
+func generatePrivateKeyBytesImpl(ctx context.Context, crt *cmapi.Certificate) ([]byte, error) {
+	signer, err := pki.GeneratePrivateKeyForCertificate(crt)
+	if err != nil {
+		return nil, err
+	}
+
+	keyData, err := pki.EncodePrivateKey(signer, crt.Spec.KeyEncoding)
+	if err != nil {
+		return nil, err
+	}
+
+	return keyData, nil
+}
+
+// secretData is a structure wrapping private key, certificate and CA data
+type secretData struct {
+	pk, cert, ca []byte
+}
+
+// setSecretValues will update the Secret resource 's' with the data contained
+// in the given secretData.
+// It will update labels and annotations on the Secret resource appropriately.
+// The Secret resource 's' must be non-nil, although may be a resource that does
+// not exist in the Kubernetes apiserver yet.
+// setSecretValues will NOT actually update the resource in the apiserver.
+// If updating an existing Secret resource returned by an api client 'lister',
+// make sure to DeepCopy the object first to avoid modifying data in-cache.
+func setSecretValues(ctx context.Context, crt *cmapi.Certificate, s *corev1.Secret, data secretData) error {
+	// initialize the `Data` field if it is nil
+	if s.Data == nil {
+		s.Data = make(map[string][]byte)
+	}
+
+	s.Data[corev1.TLSPrivateKeyKey] = data.pk
+	s.Data[corev1.TLSCertKey] = data.cert
+	s.Data[cmapi.TLSCAKey] = data.ca
+
+	if s.Annotations == nil {
+		s.Annotations = make(map[string]string)
+	}
+
+	s.Annotations[cmapi.CertificateNameKey] = crt.Name
+	s.Annotations[cmapi.IssuerNameAnnotationKey] = crt.Spec.IssuerRef.Name
+	s.Annotations[cmapi.IssuerKindAnnotationKey] = apiutil.IssuerKind(crt.Spec.IssuerRef)
+
+	// if the certificate data is empty, clear the subject related annotations
+	if len(data.cert) == 0 {
+		delete(s.Annotations, cmapi.CommonNameAnnotationKey)
+		delete(s.Annotations, cmapi.AltNamesAnnotationKey)
+		delete(s.Annotations, cmapi.IPSANAnnotationKey)
+	} else {
+		x509Cert, err := pki.DecodeX509CertificateBytes(data.cert)
+		// TODO: handle InvalidData here?
+		if err != nil {
+			return err
+		}
+
+		s.Annotations[cmapi.CommonNameAnnotationKey] = x509Cert.Subject.CommonName
+		s.Annotations[cmapi.AltNamesAnnotationKey] = strings.Join(x509Cert.DNSNames, ",")
+		s.Annotations[cmapi.IPSANAnnotationKey] = strings.Join(pki.IPAddressesToString(x509Cert.IPAddresses), ",")
+	}
+
+	return nil
 }
